ref(sdk): Add agent runtime contract
Move the Claude Code SDK execution behind a runtime adapter and expose the normalized runtime contract for future Pi support.

Refs #269
