+++ src/sdk/usage.ts
@@ -1,27 +1,5 @@
-import type { SDKResultMessage } from '@anthropic-ai/claude-agent-sdk';
 import type { UsageStats, AuxiliaryUsageMap } from '../types/index.js';
 import type { AuxiliaryUsageEntry } from './types.js';
-
-/**
- * Extract usage stats from an SDK result message.
- *
- * The Anthropic API reports `input_tokens` as only the non-cached portion.
- * We normalize so that `inputTokens` is the *total* input tokens
- * (non-cached + cache_read + cache_creation), with the cache fields
- * being subsets of that total.
- */
-export function extractUsage(result: SDKResultMessage): UsageStats {
-  const rawInput = result.usage['input_tokens'];
-  const cacheRead = result.usage['cache_read_input_tokens'] ?? 0;
-  const cacheCreation = result.usage['cache_creation_input_tokens'] ?? 0;
-  return {
-    inputTokens: rawInput + cacheRead + cacheCreation,
-    outputTokens: result.usage['output_tokens'],
-    cacheReadInputTokens: cacheRead,
-    cacheCreationInputTokens: cacheCreation,
-    costUSD: result.total_cost_usd,
-  };
-}
 
 /**
  * Create empty usage stats.
